Don't override body for all 404 stubs
This makes the test for 404 stubs with bodies in the preceding commit pass.
However it's the wrong solution to the problem, because it makes the tests
fail for un-stubbed paths with empty bodies.

I'm not sure what the correct solution to this is, but I'm hoping this
commit helps demonstrate where the problem is and can prompt a discussion
about the correct solution.